refactor: extract stat rolls and sheet UI actions (PR10)
Add stat-rolls.mjs for stat checks, sheet-rolls.mjs as roll click
router, and sheet-actions.mjs for delegated click/change handlers.
Pure species-removal updates in sheet-actions-pure.mjs with tests.
Actor sheet reduced from ~1057 to ~670 lines.
